+++ src/prelude.rs
@@ -1,3 +1,4 @@
+pub use crate::core::bus::Bus;
 pub use crate::core::message::{Composer, Message, Type as MessageType};
 pub use crate::core::persistence::reading::Reading;
 pub use crate::core::persistence::sensor::Sensor;